Refactor styles model and controller for modularity
